requirements.txt
discord.py
pytz
tinydb